Updates logic to record if no file given, then playback recorded file !!! 🔁
