+++ cli.c
@@ -1,14 +1,19 @@
 #include "uphonor.h"
 
+/* Parse the command line arguments
+
+If there are no arguments, we just start the program
+   If there is one argument, it is the file to play and loop
+   If there are two arguments, the second is the volume level. */
+
 int cli(int argc, char **argv, struct data *data)
 {
-  /* A single argument: path to the audio file. */
-  if (argc < 2)
+  if (argc > 1)
   {
-    fprintf(stderr,
-            "expected an argument: the file to open\n");
-    return 1;
+    // If there are any arguments, the first is the file to play
+    return start_playing(data, argv[1]);
   }
+
   /* If a second argument is given, it is the volume level. */
   if (argc == 3)
   {
@@ -24,15 +29,5 @@ int cli(int argc, char **argv, struct data *data)
     data->volume = 1.0f;
   }
 
-  /* We initialise libsndfile, the library we'll use to convert
-     the audio file's content into interlaced float samples. */
-  memset(&data->fileinfo, 0, sizeof(data->fileinfo));
-  data->file = sf_open(argv[1], SFM_READ, &data->fileinfo);
-  if (data->file == NULL)
-  {
-    fprintf(stderr, "file opening error: %s\n",
-            sf_strerror(NULL));
-    return 1;
-  }
   return 0;
 }
