+++ meson.build
@@ -14,4 +14,5 @@ alsa = dependency('alsa')
 executable('uphonor', 'uphonor.h', 'main.c', dependencies : [pipewire, sndfile, alsa, math], install : true)
 # executable('midi', 'midi.c', dependencies : [pipewire, alsa], install : true)
 executable('stream', 'stream.c', dependencies: [pipewire, sndfile])
+executable('play', 'play-file.c', dependencies: [pipewire, sndfile])
 executable('filter', 'filter.c', dependencies: [pipewire, math])
